==> R/07_results_integration.Rmd <==
---
title: "07_results_integration.Rmd"
author: "Naresh Jhawar"
date: "2025-04-27"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggplot2)
library(dplyr)
```

## 1. Load results
```{r load_results}
# Load the results from the previous analysis
```

## 2. Integrate results
```{r integrate_results}
# Integrate the results from different analyses
```

## 3. Visualize results
```{r visualize_results}

```

## 4. Save results
```{r save_results}
# Save the integrated results to a file
```
